requirements.txt
anthropic>=0.25.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
apscheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs>=0.5.0
lxml>=4.9.0
python-docx>=1.1.0
docx2pdf>=0.1.8
python-dotenv>=1.0.0
